requirements.txt
# Python Version 3.12.3
# Cuda Version 12.1

gymnasium
gym-notices
gymnasium
keras
mediapy
numpy
pillow
pyboy
PySDL2
pysdl2-dll
stable_baselines3
tensor
tensorboard
tensorboard-data-server
tensorboardX
tensorflow
torchvision
websockets
wheel
torchaudio
sb3_contrib
--extra-index-url https://download.pytorch.org/whl/cu121
torch